BettingRestAPI/requirements.txt
apscheduler
crawler
absl-py==0.9.0
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.1
atari-py==0.2.6
attrs==19.3.0
Automat==20.2.0
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
future==0.18.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gym==0.17.1
h5py==2.10.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.1
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
parsel==1.6.0
Pillow==7.1.2
Protego==0.1.16
protobuf==3.11.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyglet==1.5.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
Scrapy==2.1.0
scrapy-proxies==0.4
seaborn==0.10.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.0.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zope.interface==5.1.0
